+++ helpers/ordershelper.js
@@ -1,6 +1,4 @@
-const user = require('../model/userschema');
 const Product = require('../model/productschema')
-const cart = require('../model/cartschema')
 const order = require('../model/orderschema')
 module.exports = {
