+++ package.json
@@ -9,7 +9,6 @@
     "bcrypt": "^5.1.1",
     "chart.js": "^4.4.1",
     "cookie-parser": "~1.4.4",
-    "debug": "~2.6.9",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "express-fileupload": "^1.4.3",
@@ -23,10 +22,11 @@
     "mongoose": "^8.0.2",
     "morgan": "~1.9.1",
     "nodemailer": "^6.9.7",
-    "nodemon": "^3.0.2",
     "puppeteer": "^21.6.1",
     "puppeteer-html-pdf": "^3.0.3",
-    "razorpay": "^2.9.2",
-    "util": "^0.12.5"
+    "razorpay": "^2.9.2"
+  },
+  "devDependencies": {
+    "nodemon": "^3.0.2"
   }
 }
