+++ routes/admin.js
@@ -24,26 +24,25 @@ router.get('/logout',adminlogout)
 //******************************  Product Section  ********************** */
 router.get('/products', isAuth,allproducts)
 router.get('/products/add_product', isAuth,addproductpage)
-router.post('/products/add_product/:id',addproduct)
+router.post('/products/add_product/:id',isAuth,addproduct)
 router.get('/edit_product/:id', isAuth,editproduct)
 router.post('/edit_product/:id', isAuth,edit_product)
-router.get('/products/editproductview',allproducts);
-router.get('/products/deleteproduct/:id',deleteproduct);
+router.get('/products/editproductview',isAuth,allproducts);
+router.get('/products/deleteproduct/:id',isAuth,deleteproduct);
 
 //*********************  Orders  ********************** */
-
 router.get('/orders', isAuth,orders);
 router.get('/orders/confirm/:id', isAuth,confirm);
-router.get('/orders/shipped/:id',shipped);
-router.get('/orders/delivered/:id',delivered);
-router.get('/orders/cancelled/:id', cancelled);
-router.get('/ordersproduct/:id',productdetails)
-router.post('/salereport',salereport)
-router.get('/pdfgenerator',pdf);
-router.get('/gmail/:id',gmail);
+router.get('/orders/shipped/:id',isAuth,shipped);
+router.get('/orders/delivered/:id',isAuth,delivered);
+router.get('/orders/cancelled/:id',isAuth,cancelled);
+router.get('/ordersproduct/:id',isAuth,productdetails)
+router.post('/salereport',isAuth,salereport)
+router.get('/pdfgenerator',isAuth,pdf);
+router.get('/gmail/:id',isAuth,gmail);
 // Testing route
 router.get('/test',async(req,res)=>{
-  res.render('admin/invoice')
+  res.render('admin')
 })
