+++ views/users/login.hbs
@@ -2,9 +2,9 @@
 
 <head>
     <!-- Inside the head section of your login page -->
-<meta http-equiv="cache-control" content="no-cache, no-store, must-revalidate" />
-<meta http-equiv="pragma" content="no-cache" />
-<meta http-equiv="expires" content="0" />
+    <meta http-equiv="cache-control" content="no-cache, no-store, must-revalidate" />
+    <meta http-equiv="pragma" content="no-cache" />
+    <meta http-equiv="expires" content="0" />
 
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -63,7 +63,8 @@
                                             <div class="pt-1 mb-3">
                                                 <a href="/auth/google" class="btn btn-lg btn-block border border-dark">
                                                     <img src="../images/google.png" alt="Google Login"
-                                                        class="google-logo mr-2" style="height: 40px;"><span class="">Login with Google</span>
+                                                        class="google-logo mr-2" style="height: 40px;"><span
+                                                        class="">Login with Google</span>
 
                                                 </a>
                                             </div>
@@ -71,12 +72,15 @@
                                             <h5 class=" mb-3" id="errorMessage" style="color: red;">
                                                 {{errorMessage}}
                                             </h5>
-
-                                            <a class="small text-muted" href="#!">Forgot password?</a>
-                                            <p class="mb-5 pb-lg-2" style="color: #393f81;">Don't have an account? <a
-                                                    href="/users/user_registration" style="color: rgb(7, 204, 142);"
-                                                    class="strong bold">Register here</a></p>
-
+                                            <div class="text-center">
+                                                <a class="small text-muted" href="#!">Forgot password?</a>
+                                                <p class="mb-3 pb-lg-2" style="color: #393f81; font-size: 16px;">Don't
+                                                    have an account?
+                                                    <a href="/users/user_registration"
+                                                        style="color: blue; font-weight: bold; font-size: 16px;">Register
+                                                        here</a>
+                                                </p>
+                                            </div>
                                     </form>
 
                                 </div>
@@ -88,13 +92,14 @@
         </div>
     </section>
     <!-- Inside your login page -->
-{{!-- <script>
-  // Disable back button
-  history.pushState(null, null, document.URL);
-  window.addEventListener('popstate', function () {
-    history.pushState(null, null, document.URL);
-  });
-</script> --}}
+    {{!--
+    <script>
+        // Disable back button
+        history.pushState(null, null, document.URL);
+        window.addEventListener('popstate', function () {
+            history.pushState(null, null, document.URL);
+        });
+    </script> --}}
 
 </body>
